+++ gpm/main.go
@@ -37,6 +37,7 @@ var(
   DIGIT   = flag.Bool("digit", false, "use digit to generate a random password")
   LETTER  = flag.Bool("letter", false, "use letter to generate a random password")
   SPECIAL = flag.Bool("special", false, "use special chars to generate a random password")
+  EXPORT  = flag.Bool("export", false, "export a wallet in json format")
   HELP    = flag.Bool("help", false, "print this help message")
 )
 
@@ -61,5 +62,7 @@ func Run() {
     cli.updateEntry()
   } else if *DELETE {
     cli.deleteEntry()
+  } else if *EXPORT {
+    cli.ExportWallet()
   }
 }